+++ scripts/build.js
@@ -17,7 +17,6 @@ async function build() {
       entryPoints: ["src/content-script.ts", "src/popup/popup.ts", "src/options/options.ts"],
       bundle: true,
       format: "esm",
-      sourcemap: true,
       outdir: UNPACKED_OUT_DIR,
       watch: isDevMode
         ? {