+++ src/options/options.ts
@@ -1,3 +1,5 @@
+import browser from "webextension-polyfill";
+
 import { getContentString } from "../lib/github/rest-api";
 import { fitTextareaToContent } from "../lib/utils/fit-textarea-to-content";
 import { getUniqueTagsFromMarkdownString } from "../lib/utils/tags";
@@ -27,7 +29,7 @@ async function renderAllFields() {
 
 renderAllFields();
 
-chrome.storage.onChanged.addListener(function (changes, namespace) {
+browser.storage.onChanged.addListener(function (changes, namespace) {
   if (namespace === "sync") {
     renderAllFields();
   }